Fix PS3 compilation issues and finalize SDL-free support
- Resolved "identifier undefined" and "class has no member" errors by strictly guarding SDL-specific code in Drawing.cpp.
- Fixed unbalanced #if / #endif and missing braces in Drawing.cpp.
- Corrected GL_COLOR_ATTACHMENT0_OES usage and provided a fallback definition in RetroEngine.hpp.
- Replaced binary literals with hex for Visual Studio 2012 compatibility.
- Resolved sign conversion warnings and unreachable code in Collision.cpp and Debug.cpp.
- Added explicit casts for float-to-int/short conversions to silence warnings.
- Finalized PSGL mappings and platform-specific configurations for PS3.

+++ RSDKv3/Palette.hpp
@@ -69,9 +69,9 @@ inline void SetActivePalette(byte newActivePal, int startLine, int endLine)
     }
 }
 
-inline void SetPaletteEntry(byte paletteIndex, byte index, byte r, byte g, byte b)
+inline void SetPaletteEntry(sbyte paletteIndex, byte index, byte r, byte g, byte b)
 {
-    if (paletteIndex != 0xFF) {
+    if (paletteIndex != (sbyte)0xFF) {
         PACK_RGB888(fullPalette[paletteIndex][index], r, g, b);
         fullPalette32[paletteIndex][index].r = r;
         fullPalette32[paletteIndex][index].g = g;
